refactor(banking): move settlement checks out of TransactionFingerprint
Asking a class that builds a hash whether a transaction has settled reads
wrong, and it was only there because that is where the card-code pair
already lived.

TransactionSettlement now owns both signals: the `status` list that gates
the import, and the MCRD/UPCT -> CCRD/POSD canonicalization that feeds
the content hash. TransactionFingerprint keeps only what identifies a
transaction - UNSTABLE_ID_BANKS and the hashing itself - and asks
TransactionSettlement for the card code.

No behaviour change.
